+++ Utilities/Loaders/ImageLoader.h
@@ -2,8 +2,6 @@
 
 #include <Magick++.h>
 
-#include <cuda_fp16.h>
-
 #include <assert.h>
 #include <math.h>
 
@@ -21,7 +19,7 @@ class ImageLoader {
         double minAspectRatio, double maxAspectRatio, uint32_t outputImageRows, uint32_t outputImageColumns, Magick::Image &image);
 
     // Returns 1 byte for each of rgb in CHW format, for example [r0, g0, b0, r1, g1, b1, ...]
-    static bool toRgbArray(Magick::Image &image, half *rgbPixelArray, bool toHWCLayout);
+    static bool toRgbArray(Magick::Image &image, uint8_t *rgbPixelArray, bool toHWCLayout);
 
    private:
     ImageLoader() {}
